@@ -1,19 +1,39 @@
 apply plugin: 'com.android.library'
-apply plugin: 'com.jakewharton.butterknife'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-android-extensions'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.2"
+    compileSdkVersion 29
+    buildToolsVersion "29.0.3"
 
     defaultConfig {
-        minSdkVersion 15
-        targetSdkVersion 25
-        versionCode 9
-        versionName "1.0.8"
+        minSdkVersion 21
+        targetSdkVersion 29
+        versionCode 12
+        versionName "1.0.12"
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
 
+        packagingOptions {
+            exclude "lib/mips/*.so"
+        }
     }
+
+    signingConfigs {
+        debug {
+            keyAlias 'chipper'
+            keyPassword 'chipper.2019'
+            storeFile file('chipper.keystore')
+            storePassword 'chipper.2019'
+        }
+        release {
+            keyAlias 'chipper'
+            keyPassword 'chipper.2019'
+            storeFile file('chipper.keystore')
+            storePassword 'chipper.2019'
+        }
+    }
+
     buildTypes {
         release {
             minifyEnabled false
@@ -24,23 +44,24 @@ android {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
+    androidTestCompile('androidx.test.espresso:espresso-core:3.1.0', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
-    compile 'com.android.support:appcompat-v7:25.3.0'
+    compile 'androidx.appcompat:appcompat:1.1.0'
     testCompile 'junit:junit:4.12'
 
     // recycler
-    compile 'com.android.support:recyclerview-v7:25.3.0'
+    compile 'androidx.recyclerview:recyclerview:1.1.0'
     compile 'com.beloo.widget:ChipsLayoutManager:0.3.7@aar'
 
     // circle image view
-    compile 'de.hdodenhof:circleimageview:2.1.0'
-
-    // butter knife
-    compile 'com.jakewharton:butterknife:8.5.1'
-    annotationProcessor 'com.jakewharton:butterknife-compiler:8.5.1'
+    compile 'de.hdodenhof:circleimageview:2.2.0'
+    implementation "androidx.core:core-ktx:+"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 }
 
 apply plugin: 'com.github.dcendents.android-maven'
 group='com.github.pchmn'
+repositories {
+    mavenCentral()
+}
