@@ -3,8 +3,8 @@ package com.pchmn.materialchips.model;
 
 import android.graphics.drawable.Drawable;
 import android.net.Uri;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 public class Chip implements ChipInterface {
 
